#189 revisited celery_worker lifecycle which was being invoked more than once
